Changes: Added prediction methods based on velocity calculations to update feet and hip positions when kinect prediction has not processed yet. This decreased the latency of updating the trackers in space (made the trackers track quicker movements better). Right now only positions are updated. Velocity stays constant for performance until the impact of adding this can be benchmarked.
Reduced estimated time differences in velocity calculation to accomodate reduced update latency

Added velocity calculation for hip tracker

Changed the way the vector position was updated. Instead of using just the raw data, the raw data is merged with the predicted position (same as the prediction method above) at a ratio of 6 to 4. This decreased the variability of the data (made the trackers less jittery) for the feet to a level I found usable.

Removed temporal smoothing tweaks in the calibrator as any level of smoothing set above 0.05 resulted in trackers that were too slow to use.

In the future, (maybe if more people want this functionality) timing data in the function needs to be changed to be dynamically computed instead of hardcoded estimations (estimations based on benchmarks found here https://github.com/microsoft/Azure-Kinect-Sensor-SDK/issues/514). This needs to be done to accomadate different machine setups and to also support newer software/hardware that can process frames at a higher fps.

TODOs :
1: figure out a way to preserve hip positional data when rotating body ( when body is rotated the hip's predicted position becomes whatever side is facing the camera. This breaks some of the rotational tracking capabilities of the device. Potentially look into setting the hip tracker on one or some combination of the hip bones instead of the pelvis bone?

2: Measure the performance impact of predicting data vs using only raw data

3: Implement the method for more trackers

4: Add accurate timing data into the prediction method so the predicted values can be more accurate for every machine (not just my testing rig).

Testing Rig Specs:
i7-8700k
MSI 370 A-Pro Motherboard
32gb ddr4 ram
rx 580 8gb
gtx 960 2gb
Rift S
Kinect Azure
Windows 10
All drivers and software up to date
7/26/2020